띄어쓰기

+++ src/Encoding.cc
@@ -1,12 +1,14 @@
 #include "Encoder.hh"
 
+#include <cstdint>
+
 namespace Dlink
 {
 	static bool is_surrogate(char16_t c)
 	{
-		return c - 0xD800 < 2048;
+		return static_cast<std::int32_t>(c) - 0xD800 < 2048;
 	}
-	static bool is_high_surrogate (char16_t c)
+	static bool is_high_surrogate(char16_t c)
 	{
 		return c & 0xFFFFFC00 == 0xD800;
 	}
